+++ src/MVDegrains.cpp
@@ -163,7 +163,7 @@ static const VSFrameRef *VS_CC mvdegrainGetFrame(int n, int activationReason, vo
             const VSFrameRef *frame = vsapi->getFrameFilter(n, d->vectors[r], frameCtx);
             fgopInit(&fgops[r], &d->vectors_data[r]);
             const VSMap *mvprops = vsapi->getFramePropsRO(frame);
-            fgopUpdate(&fgops[r], (const int *)vsapi->propGetData(mvprops, prop_MVTools_vectors, 0, NULL));
+            fgopUpdate(&fgops[r], (const uint8_t *)vsapi->propGetData(mvprops, prop_MVTools_vectors, 0, NULL));
             isUsable[r] = fgopIsUsable(&fgops[r], d->nSCD1, d->nSCD2);
             vsapi->freeFrame(frame);
