+++ src/MVMask.c
@@ -107,7 +107,7 @@ static const VSFrameRef *VS_CC mvmaskGetFrame(int n, int activationReason, void 
         const VSFrameRef *mvn = vsapi->getFrameFilter(n, d->vectors, frameCtx);
         fgopInit(&fgop, &d->vectors_data);
         const VSMap *mvprops = vsapi->getFramePropsRO(mvn);
-        fgopUpdate(&fgop, (const int *)vsapi->propGetData(mvprops, prop_MVTools_vectors, 0, NULL));
+        fgopUpdate(&fgop, (const uint8_t *)vsapi->propGetData(mvprops, prop_MVTools_vectors, 0, NULL));
         vsapi->freeFrame(mvn);
 
         const int kind = d->kind;
