+++ src/MVSCDetection.c
@@ -64,7 +64,7 @@ static const VSFrameRef *VS_CC mvscdetectionGetFrame(int n, int activationReason
         FakeGroupOfPlanes fgop;
         fgopInit(&fgop, &d->vectors_data);
         const VSMap *mvprops = vsapi->getFramePropsRO(mvn);
-        fgopUpdate(&fgop, (const int *)vsapi->propGetData(mvprops, prop_MVTools_vectors, 0, NULL));
+        fgopUpdate(&fgop, (const uint8_t *)vsapi->propGetData(mvprops, prop_MVTools_vectors, 0, NULL));
         vsapi->freeFrame(mvn);
 
         const char *propNames[2] = { "_SceneChangePrev", "_SceneChangeNext" };
